+++ src/test/scala/io/github/bcarter97/GeneratorSpec.scala
@@ -3,110 +3,77 @@ package io.github.bcarter97
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
 
-class GeneratorSpec extends AnyWordSpecLike with Matchers {
+class GeneratorV2Spec extends AnyWordSpecLike with Matchers {
 
-  "Generator" should {
-    "throw an error if Generator is called with an invalid maxIndex value" in {
-      assertThrows[IllegalArgumentException] {
-        Generator(0)
-      }
+  "GeneratorV2" should {
+    "Generate a reproducible PrimaryId" in {
+      val generatorV2 = GeneratorV2(maxIndex = 100)
+      val id          = generatorV2.primaryId(50)
+      val id2         = generatorV2.primaryId(50)
+
+      id shouldBe id2
     }
 
-    "throw an error if Generator is called with an invalid subId value" in {
-      assertThrows[IllegalArgumentException] {
-        Generator(1, 0)
-      }
-    }
+    "Generate a range of reproducible PrimaryIds" in {
+      val generator = GeneratorV2(maxIndex = 10)
+      val ids       = generator.primaryIds(5, 10)
+      val ids2      = generator.primaryIds(5, 10)
 
-    "generate a reproducible UUID" in {
-      val generator = Generator(maxIndex = 100)
-      val id        = generator.id(50)
-      val id2       = generator.id(50)
-
-      id2 shouldBe id
-    }
-
-    "generate a range of reproducible UUIDs" in {
-      val generator = Generator(maxIndex = 10)
-      val ids       = generator.ids(5, 7)
-      val ids2      = generator.ids(5, 7)
-
-      ids.length shouldBe 3
-      ids2 shouldBe ids
+      ids shouldBe ids2
     }
 
     "guard against going out of bounds when generating a range of reproducible UUIDs" in {
-      val generator = Generator(maxIndex = 10)
-      val ids       = generator.ids(-1, 11)
-      val ids2      = generator.ids(-1, 11)
+      val generator = GeneratorV2(maxIndex = 10)
+      val ids       = generator.primaryIds(-1, 11)
+      val ids2      = generator.primaryIds(-1, 11)
 
       ids.length shouldBe 10
       ids2 shouldBe ids
     }
 
     "generate a reproducible amount of subIds given an id" in {
-      val generator  = Generator(maxIndex = 100)
-      val id         = generator.id(50)
-      val id2        = generator.id(50)
-      val numSubIds  = generator.numSubIdsForId(id)
-      val numSubIds2 = generator.numSubIdsForId(id2)
+      val generator = GeneratorV2(maxIndex = 100)
+      val id        = generator.primaryId(50)
+      val id2       = generator.primaryId(50)
 
-      numSubIds2 shouldBe numSubIds
+      id.numSubIds shouldBe id2.numSubIds
     }
 
     "generate a reproducible list of subIds given an id" in {
-      val generator = Generator(maxIndex = 100)
-      val id        = generator.id(50)
-      val id2       = generator.id(50)
-      val subIds    = generator.subIdsFromId(id)
-      val subIds2   = generator.subIdsFromId(id2)
+      val generator = GeneratorV2(maxIndex = 100)
+      val id        = generator.primaryId(50)
+      val id2       = generator.primaryId(50)
 
-      subIds shouldBe subIds2
+      id.subIds shouldBe id2.subIds
     }
 
-    "return which index created a specific id" in {
-      val generator = Generator(maxIndex = 100)
-      val id        = generator.id(50)
-      val idIndex   = generator.index(id)
+    "return which index created a specific primaryId" in {
+      val generator = GeneratorV2(maxIndex = 100)
+      val id        = generator.primaryId(50)
+      generator.clearCache()
 
-      idIndex shouldBe Some(50)
+      generator.index(id.uuid) shouldBe Some(50)
     }
 
     "return none if no index for a specific id was found" in {
-      val generator = Generator(maxIndex = 100)
-      val id        = generator.id(500)
-      val idIndex   = generator.index(id)
+      val generator = GeneratorV2(maxIndex = 100)
+      val id        = generator.primaryId(500)
+      generator.clearCache()
 
-      idIndex shouldBe None
-    }
-
-    "find if a subId is part of an id's sub group" in {
-      val generator = Generator(maxIndex = 100)
-      val id        = generator.id(50)
-      val subIds    = generator.subIdsFromId(id)
-
-      subIds.foreach(subId => generator.isSubIdFromId(subId, id) shouldBe true)
+      generator.index(id.uuid) shouldBe None
     }
 
     "find the id that generated a list of subIds" in {
-      val generator = Generator(maxIndex = 100)
-      val id        = generator.id(50)
-      val subIds    = generator.subIdsFromId(id)
+      val generator = GeneratorV2(maxIndex = 100)
+      val id        = generator.primaryId(50)
+      val subIds    = id.subIds
+      generator.clearCache()
 
-      subIds.foreach(subId => generator.idFromSubId(subId) shouldBe Some(id))
-    }
-
-    "find the index that generated a list of subIds" in {
-      val generator = Generator(maxIndex = 100)
-      val index     = 50
-      val id        = generator.id(index)
-      val subIds    = generator.subIdsFromId(id)
-
-      subIds.foreach(subId => generator.indexFromSubId(subId) shouldBe Some(50))
+      subIds.foreach(subId => generator.subIdFromUuid(subId.uuid).map(_.primaryId) shouldBe Some(id))
     }
 
     "sample a range of ids" in {
-      val generator = Generator(maxIndex = 100)
+      val generator = GeneratorV2(maxIndex = 100)
       val ids       = generator.sample(10)
 
       ids.length shouldBe 10
@@ -114,7 +81,7 @@ class GeneratorSpec extends AnyWordSpecLike with Matchers {
     }
 
     "sample a single id if no parameter is specified" in {
-      val generator = Generator(maxIndex = 100)
+      val generator = GeneratorV2(maxIndex = 100)
       val id        = generator.sample()
       val id2       = generator.sample()
 
@@ -123,18 +90,18 @@ class GeneratorSpec extends AnyWordSpecLike with Matchers {
 
     "throw an error if sample is called with a number < 1" in {
       assertThrows[IllegalArgumentException] {
-        Generator(1).sample(0)
+        GeneratorV2(1).sample(0)
       }
     }
 
     "call sample if no index is passed to id" in {
-      val generator  = Generator(maxIndex = 100)
-      val generator2 = Generator(maxIndex = 100)
+      val generator  = GeneratorV2(maxIndex = 100)
+      val generator2 = GeneratorV2(maxIndex = 100)
 
-      val id      = generator.id()
-      val nextId  = generator.id()
-      val id2     = generator2.id()
-      val nextId2 = generator2.id()
+      val id      = generator.primaryId()
+      val nextId  = generator.primaryId()
+      val id2     = generator2.primaryId()
+      val nextId2 = generator2.primaryId()
 
       id should not be nextId
       id shouldBe id2
@@ -142,7 +109,7 @@ class GeneratorSpec extends AnyWordSpecLike with Matchers {
     }
 
     "return two distinct ranges of ids" in {
-      val generator = Generator(maxIndex = 100)
+      val generator = GeneratorV2(maxIndex = 100)
       val ids       = generator.sample(10)
       val ids2      = generator.sample(10)
 
@@ -152,7 +119,7 @@ class GeneratorSpec extends AnyWordSpecLike with Matchers {
     }
 
     "loop round to the original ids if the max index is hit" in {
-      val generator = Generator(maxIndex = 10)
+      val generator = GeneratorV2(maxIndex = 10)
       val ids       = generator.sample(10)
       val ids2      = generator.sample(10)
